// frontend/src/modules/examination-platform/enums/ExamStatus.ts
export enum ExamStatus {
  DRAFT = 'draft',
  PENDING_APPROVAL = 'pending_approval',
  APPROVED = 'approved',
  SCHEDULED = 'scheduled',
  HALL_TICKETS_GENERATED = 'hall_tickets_generated',
  EXAM_ACTIVE = 'exam_active',
  EXAM_CLOSED = 'exam_closed',
  EVALUATION = 'evaluation',
  RESULTS_PUBLISHED = 'results_published',
  ARCHIVED = 'archived',
}